move solution transformation to new file
